==> 1_data_collection.ipynb <==
# Project 3 NLP of Subreddits

The goal of this project is to take two subreddits and train several NLP models to differentiate between the two and predict which submissions and comments are from each subreddit. I chose two subreddits of podcast networks, Gimlet and Maximum Fun. I chose these two because there would be some overlap of vocabulary because they are both podcast networks, but they would be different enough because the shows of each network cover different kinds of topics. 

Maximum Fun was started in 2004 by Jesse Thorn, originally an NPR host. Most of the podcasts are comedy and culture themed. I got to know Maximum Fun because of the podcast Judge John Hodgman, in which Hodgman presides over a fake internet court and settles petty disputes between friends and family members. 

Gimlet was started in 2014 by another NPR host, Alex Blumberg. The content of the podcasts tends to be more serious than Maximum Fun, and include a healthy dose of true crime content. My favorite show from the network is Heavy Weight, where host Jonathan Goldstein helps people confront regrets and defining moments in the lives of his guests.

### Contents:
- [Functions for Collecting Subreddits](#Functions-for-Collecting-Subreddits)
- [Pulling Maximum Fun Subreddit Submissions](#Pulling-Maximum-Fun-Subreddit-Submissions)
- [Pulling Maximum Fun Subreddit Comments](#Pulling-Maximum-Fun-Subreddit-Comments)
- [Pulling Gimlet Subreddit Submissions](#Pulling-Gimlet-Subreddit-Submissions)
- [Pulling Gimlet Subreddit Comments](#Pulling-Maximum-Fun-Subreddit-Comments)

# importing packages 
import requests
import pandas as pd

## Functions for Collecting Subreddits

# Function for getting reddit submissions 
def get_reddit_submissions(subreddit, UTC):
    # URL for pushshift API
    url = 'https://api.pushshift.io/reddit/search/submission'
    # Parameters for pulling the data
    params = {
        'subreddit': subreddit,
        'size': 500, 
        'before': UTC
    }
    # Creating the request
    response = requests.get(url, params)
    # Grabbing the data
    data = response.json()
    # Creating a variable called posts to store data
    posts = data['data']
    # Storing data in a dataframe 
    return pd.DataFrame(posts)

# Function for getting reddit comments 
def get_reddit_comments(subreddit, UTC):
    # URL for pushshift API
    url = 'https://api.pushshift.io/reddit/search/comment'
    # Parameters for pulling the data
    params = {
        'subreddit': subreddit,
        'size': 500, 
        'before': UTC
    }
    # Creating the request
    response = requests.get(url, params)
    # Grabbing the data
    data = response.json()
    # Creating a variable called posts to store data
    posts = data['data']
    # Storing data in a dataframe 
    return pd.DataFrame(posts)

## Pulling Maximum Fun Subreddit Submissions 

# For the first time calling this function I left out the UTC and altered the function 
max_fun_sub1 = get_reddit_submissions('maximumfun')

# Checking the data
max_fun_sub1.head()

# I chose the following columns as ones that would be useful because of the content they contain
max_fun_sub1 = max_fun_sub1[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

# Checking the shape of the data
max_fun_sub1.shape

# Checking the data types to see what they are
max_fun_sub1.info()

# Getting the minimum values for the UTC
max_fun_sub1.min()

# Pulling on the minimum UTC from before
max_fun_sub2 = get_reddit_submissions('maximumfun', 1569869169)

### From here the function is run for 4 more iterations

max_fun_sub2 = max_fun_sub2[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

max_fun_sub2.min()

max_fun_sub3 = get_reddit_submissions('maximumfun', 1559240433)

max_fun_sub3 = max_fun_sub3[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

max_fun_sub3.min()

max_fun_sub4 = get_reddit_submissions('maximumfun', 1551117329)

max_fun_sub4 = max_fun_sub4[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

max_fun_sub4.min()

max_fun_sub5 = get_reddit_submissions('maximumfun', 1540668669)

max_fun_sub5 = max_fun_sub5[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

max_fun_sub5.min()

## Pulling Maximum Fun Subreddit Comments

### I decided to get both submissions and comments from each of my subreddits. The comments for the subreddits have a lot more content, particularly for Maximum Fun, where a lot of the submissions are links to episodes. 

# First pull for comments, and this is repeated 15 times. 
max_fun_comments1 = get_reddit_comments('maximumfun')

# These are the columns I found to be useful for content
max_fun_comments1 = max_fun_comments1[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments1.min()

max_fun_comments2 = get_reddit_comments('maximumfun', 1578596547)

max_fun_comments2 = max_fun_comments2[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments2.min()

max_fun_comments3 = get_reddit_comments('maximumfun', 1576668608)

max_fun_comments3 = max_fun_comments3[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments3.min()

max_fun_comments4 = get_reddit_comments('maximumfun', 1575649111)

max_fun_comments4 = max_fun_comments4[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments4.min()

max_fun_comments5 = get_reddit_comments('maximumfun', 1574368606)

max_fun_comments5 = max_fun_comments5[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments5.min()

max_fun_comments6 = get_reddit_comments('maximumfun', 1572888504)

max_fun_comments6 = max_fun_comments6[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments6.min()

max_fun_comments7 = get_reddit_comments('maximumfun', 1571354268)

max_fun_comments7 = max_fun_comments7[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments7.min()

max_fun_comments8 = get_reddit_comments('maximumfun', 1569871527)

max_fun_comments8 = max_fun_comments8[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments8.min()

max_fun_comments9 = get_reddit_comments('maximumfun', 1567951674)

max_fun_comments9 = max_fun_comments9[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments9.min()

max_fun_comments10 = get_reddit_comments('maximumfun', 1565720234)

max_fun_comments10 = max_fun_comments10[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments10.min()

max_fun_comments11 = get_reddit_comments('maximumfun', 1564097606)

max_fun_comments11 = max_fun_comments11[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments11.min()

max_fun_comments12 = get_reddit_comments('maximumfun', 1561893862)

max_fun_comments12 = max_fun_comments12[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments12.min()

max_fun_comments13 = get_reddit_comments('maximumfun', 1560285016)

max_fun_comments13 = max_fun_comments13[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments13.min()

max_fun_comments14 = get_reddit_comments('maximumfun', 1558707053)

max_fun_comments14 = max_fun_comments14[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments14.min()

max_fun_comments15 = get_reddit_comments('maximumfun', 1557153832)

max_fun_comments15 = max_fun_comments15[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments15.min()

# I did a concat for the submission pulls I did. I only did one pull for submissions because they went 
# much further back than comments, because there are much fewer of them
max_fun_subs = pd.concat([max_fun_sub1, max_fun_sub2, max_fun_sub3, max_fun_sub4, max_fun_sub5], sort=True)

max_fun_subs.shape

# I came back and did a few different pulls to get more content, this is the first set concatenated 
max_fun_comments = pd.concat([max_fun_comments1, max_fun_comments2, max_fun_comments3, max_fun_comments4, max_fun_comments5, max_fun_comments6, max_fun_comments7, max_fun_comments8, max_fun_comments9, max_fun_comments10], sort=True)

# This was the second set of comments I pulled and concatenated 
max_fun_comments2 = pd.concat([max_fun_comments11, max_fun_comments12, max_fun_comments13, max_fun_comments14, max_fun_comments15], sort=True)

# Looking at the shape of the new comments file 
max_fun_comments.shape

# Saving the file to a csv
max_fun_subs.to_csv('./data/max_fun_submissions.csv', index=False)

# Saving the file to a csv
max_fun_comments.to_csv('./data/max_fun_comments.csv', index=False)

# Saving the file to a csv 
max_fun_comments2.to_csv('./data/max_fun_comments2.csv', index=False)

## Pulling Gimlet Subreddit Submissions

# Repeating the same process with Gimlet that I did with Maximum Fun
gimlet_sub1 = get_reddit_submissions('gimlet')

gimlet_sub1 = gimlet_sub1[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

gimlet_sub1.min()

gimlet_sub2 = get_reddit_submissions('gimlet', 1558987823)

gimlet_sub2 = gimlet_sub2[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

gimlet_sub2.min()

gimlet_sub3 = get_reddit_submissions('gimlet', 1540886253)

gimlet_sub3 = gimlet_sub3[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

gimlet_sub3.min()

gimlet_sub4 = get_reddit_submissions('gimlet', 1520011356)

gimlet_sub4 = gimlet_sub4[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

gimlet_sub4.min()

gimlet_sub5 = get_reddit_submissions('gimlet', 1496154847)

gimlet_sub5 = gimlet_sub5[['subreddit', 'author', 'selftext', 'title', 'created_utc']]

gimlet_sub5.min()

### Pulling Gimlet Subreddit Comments

# Repeating the same process for Subreddit Comments as above 
gimlet_com1 = get_reddit_comments('gimlet')

gimlet_com1 = gimlet_com1[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com1.min()

gimlet_com2 = get_reddit_comments('gimlet', 1577165162)

gimlet_com2 = gimlet_com2[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com2.min()

gimlet_com3 = get_reddit_comments('gimlet', 1574999373)

gimlet_com3 = gimlet_com3[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com3.min()

gimlet_com4 = get_reddit_comments('gimlet', 1573314112)

gimlet_com4 = gimlet_com4[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com4.min()

gimlet_com5 = get_reddit_comments('gimlet', 1571762316)

gimlet_com5 = gimlet_com5[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com5.min()

gimlet_com6 = get_reddit_comments('gimlet', 1570215341)

gimlet_com6 = gimlet_com6[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com6.min()

gimlet_com7 = get_reddit_comments('gimlet', 1567931927)

gimlet_com7 = gimlet_com7[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com7.min()

gimlet_com8 = get_reddit_comments('gimlet', 1565337684)

gimlet_com8 = gimlet_com8[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com8.min()

gimlet_com9 = get_reddit_comments('gimlet', 1563200750)

gimlet_com9 = gimlet_com9[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com9.min()

gimlet_com10 = get_reddit_comments('gimlet', 1561660350)

gimlet_com10 = gimlet_com10[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com10.min()

gimlet_com11 = get_reddit_comments('gimlet', 1559631786)

gimlet_com11 = gimlet_com11[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com11.min()

gimlet_com12 = get_reddit_comments('gimlet', 1557420807)

gimlet_com12 = gimlet_com12[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com12.min()

gimlet_com13 = get_reddit_comments('gimlet', 1555810652)

gimlet_com13 = gimlet_com13[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com13.min()

gimlet_com14 = get_reddit_comments('gimlet', 1555007585)

gimlet_com14 = gimlet_com14[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com14.min()

gimlet_com15 = get_reddit_comments('gimlet', 1553881779)

gimlet_com15 = gimlet_com15[['subreddit', 'author', 'body', 'created_utc']]

gimlet_com15.min()

# Concatenating the gimlet submissions 
gimlet_submissions = pd.concat([gimlet_sub1, gimlet_sub2, gimlet_sub3, gimlet_sub4, gimlet_sub5], sort=True)

gimlet_submissions.shape

gimlet_submissions.head()

# Concatenating the first set of gimlet comments 
gimlet_comments = pd.concat([gimlet_com1, gimlet_com2, gimlet_com3, gimlet_com4, gimlet_com5, gimlet_com6, gimlet_com7, gimlet_com8, gimlet_com9, gimlet_com10], sort=True)

gimlet_comments.shape

gimlet_comments.head()

# Concatenating the second set of Gimlet Comments
gimlet_comments2 = pd.concat([gimlet_com11, gimlet_com12, gimlet_com13, gimlet_com14, gimlet_com15], sort=True)

# Saving the Gimlet submissions to a csv
gimlet_submissions.to_csv('./data/gimlet_submissions.csv', index=False)

# Saving the first set of Gimlet comments to a csv
gimlet_comments.to_csv('./data/gimlet_comments.csv', index=False)

gimlet_comments2.head()

# Saving the second set of Gimlet comments to a csv 
gimlet_comments2.to_csv('./data/gimlet_comments2.csv', index=False)
