==> 1.5_data_collection2.ipynb <==
# Data Collection, Part II

### I decided to follow the same process as in the previous notebook to collect 5000 more comments for each subreddit. I decided to do more comments because I had submissions that went much further back than the comments I had. This is again just following the same process I did for the previous notebook. I ran the function for another 10 times and saved the resulting dataframes to csv. 

import requests
import pandas as pd

def get_reddit_submissions(subreddit, UTC):
    
    url = 'https://api.pushshift.io/reddit/search/submission'
    
    params = {
        'subreddit': subreddit,
        'size': 500, 
        'before': UTC
    }

    response = requests.get(url, params)

    data = response.json()

    posts = data['data']

    return pd.DataFrame(posts)

def get_reddit_comments(subreddit, UTC):
    
    url = 'https://api.pushshift.io/reddit/search/comment'

    params = {
        'subreddit': subreddit,
        'size': 500, 
        'before': UTC
    }

    response = requests.get(url, params)

    data = response.json()

    posts = data['data']

    return pd.DataFrame(posts)

max_fun_comments16 = get_reddit_comments('maximumfun', 1556127743)

max_fun_comments16 = max_fun_comments16[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments16.min()

max_fun_comments17 = get_reddit_comments('maximumfun', 1554578726)

max_fun_comments17 = max_fun_comments17[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments17.min()

max_fun_comments18 = get_reddit_comments('maximumfun', 1553822690)

max_fun_comments18 = max_fun_comments18[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments18.min()

max_fun_comments19 = get_reddit_comments('maximumfun', 1553116769)

max_fun_comments19 = max_fun_comments19[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments19.min()

max_fun_comments20 = get_reddit_comments('maximumfun', 1552074827)

max_fun_comments20 = max_fun_comments20[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments20.min()

max_fun_comments21 = get_reddit_comments('maximumfun', 1550256210)

max_fun_comments21 = max_fun_comments21[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments21.min()

max_fun_comments22 = get_reddit_comments('maximumfun', 1548816469)

max_fun_comments22 = max_fun_comments22[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments22.min()

max_fun_comments23 = get_reddit_comments('maximumfun', 1547056418)

max_fun_comments23 = max_fun_comments23[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments23.min()

max_fun_comments24 = get_reddit_comments('maximumfun', 1545198153)

max_fun_comments24 = max_fun_comments24[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments24.min()

max_fun_comments25 = get_reddit_comments('maximumfun', 1543833961)

max_fun_comments25 = max_fun_comments25[['subreddit', 'author', 'body', 'created_utc']]

max_fun_comments25.min()

gimlet_comments16 = get_reddit_comments('gimlet', 1553281427)

gimlet_comments16 = gimlet_comments16[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments16.min()

gimlet_comments17 = get_reddit_comments('gimlet', 1552315586)

gimlet_comments17 = gimlet_comments17[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments17.min()

gimlet_comments18 = get_reddit_comments('gimlet', 1550499154)

gimlet_comments18 = gimlet_comments18[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments18.min()

gimlet_comments19 = get_reddit_comments('gimlet', 1549582229)

gimlet_comments19 = gimlet_comments19[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments19.min()

gimlet_comments20 = get_reddit_comments('gimlet', 1549295316)

gimlet_comments20 = gimlet_comments20[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments20.min()

gimlet_comments21 = get_reddit_comments('gimlet', 1548775956)

gimlet_comments21 = gimlet_comments21[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments21.min()

gimlet_comments22 = get_reddit_comments('gimlet', 1546873218)

gimlet_comments22 = gimlet_comments22[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments22.min()

gimlet_comments23 = get_reddit_comments('gimlet', 1545330790)

gimlet_comments23 = gimlet_comments23[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments23.min()

gimlet_comments24 = get_reddit_comments('gimlet', 1544393935)

gimlet_comments24 = gimlet_comments24[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments24.min()

gimlet_comments25 = get_reddit_comments('gimlet', 1543585314)

gimlet_comments25 = gimlet_comments25[['subreddit', 'author', 'body', 'created_utc']]

gimlet_comments25.min()

max_fun_comments3 = pd.concat([max_fun_comments16, max_fun_comments17, max_fun_comments18, max_fun_comments19, max_fun_comments20, max_fun_comments21, max_fun_comments22, max_fun_comments23, max_fun_comments24, max_fun_comments25], sort=True)

max_fun_comments3.to_csv('./data/max_fun_comments3.csv', index=False)

gimlet_comments3 = pd.concat([gimlet_comments16, gimlet_comments17, gimlet_comments18, gimlet_comments19, gimlet_comments20, gimlet_comments21, gimlet_comments22, gimlet_comments23, gimlet_comments24, gimlet_comments25], sort=True)

gimlet_comments3.to_csv('./data/gimlet_comments3.csv', index=False)
