==> 3_preprocessing_.ipynb <==
# Preprocessing for NLP

# importing libraries for text pre-processing
import pandas as pd
import nltk
import string 
from nltk.stem import WordNetLemmatizer

# Reading in my cleaned reddits csv
reddits = pd.read_csv('./data/reddits.csv')

# Checking out the data
reddits.head()

# Removing the punctuation from my columns I plan on using
reddits['author'] = reddits['author'].str.replace('[^\w\s]','')
reddits['title'] = reddits['title'].str.replace('[^\w\s]','')
reddits['body'] = reddits['body'].str.replace('[^\w\s]','')
reddits['selftext'] = reddits['selftext'].str.replace('[^\w\s]','')

# Making sure it worked
reddits.head()

# Making everything lower case
reddits['author'] = reddits['author'].str.lower() 
reddits['title'] = reddits['title'].str.lower() 
reddits['body'] = reddits['body'].str.lower() 
reddits['selftext'] = reddits['selftext'].str.lower() 

# Checking out my dataframe
reddits.head()

# Making a copy of reddits so I can have one dataframe with authors and one without 
reddits2 = reddits.copy()

# Checking reddits2
reddits2.head()

# Making a words column, and reddits2 will have author, title, body, and selftext
reddits2['words'] = reddits2['author'] + ' ' + reddits2['title'] + ' ' + reddits2['body'] + ' ' + reddits2['selftext']

# Making a words column for reddits, which will have everything above other than the author
reddits['words'] = reddits['title'] + ' ' + reddits['body'] + ' ' + reddits['selftext']

# Looking at reddits
reddits.head()

# Looking at reddits2
reddits2.head()

# Dropping the columns I combined or don't need
reddits.drop(columns=['author', 'title', 'body', 'selftext'], inplace=True)
reddits2.drop(columns=['author', 'title', 'body', 'selftext'], inplace=True)

# Checking out reddits
reddits.head()

# Checking out reddits 2
reddits2.head()

# Replacing the underscore I used for nulls with a space
reddits['words'] = reddits['words'].str.replace('_', ' ')
reddits2['words'] = reddits2['words'].str.replace('_', ' ')

# Looking at reddits
reddits.head()

# Making Maximum Fun my positive class
reddits['subreddit'].replace('maximumfun', 1, inplace=True)
reddits2['subreddit'].replace('maximumfun', 1, inplace=True)

# Making Gimlet my negative class
reddits['subreddit'].replace('gimlet', 0, inplace=True)
reddits2['subreddit'].replace('gimlet', 0, inplace=True)

# I tested out lemmatization and it did not quite give me the results I expected. This is something I would
# like to experiment further with this
lemmatizer = nltk.stem.WordNetLemmatizer()

def lemmatize(words):
    items = []
    lemmas = [lemmatizer.lemmatize(x) for x in (words)]
    items.append(lemmas)
    return items
lemmatize(reddits['words'][:5])

# Looking at the info for reddits
reddits.info()

# Looking at the info for reddits2
reddits2.info()

# Reading my dataframes into two files, one with authors and one without
# I will use my data without authors to test my models, but then look at my best models and compare with
# authors included
reddits.to_csv('./data/reddits_preprocessed.csv', index=False)
reddits2.to_csv('./data/reddits_preprocessed_authors.csv', index=False)
